Renames according to discussion with MFisher
